Add Intel XPU support
Make tritonbench runnable on Intel GPUs and get `test/test_gpu` green on XPU.

Device abstraction
- `env_utils`: add `get_current_device()` (built on
  `torch.accelerator.current_accelerator()`, falls back to "cuda"),
  `get_device_module()` and `get_num_sms()` (XPU reports
  `gpu_subslice_count` instead of `multi_processor_count`).
- Replace the per-file `device="cuda"` / `torch.cuda.*` / open-coded NUM_SMS
  branches across kernels and operators with those shared helpers, so device
  dispatch lives in one place instead of being duplicated per operator.
- `gpu_utils.get_gpu_device_name()` and `run_utils` env checks understand xpu;
  `parser` accepts `--device xpu`.
- Rename `estimate_cuda_runtime_ms` to `estimate_gpu_runtime_ms`, keeping the
  old name as an alias for backward compatibility.
- `test/test_gpu/main.py` derives `TEST_DEVICE` from the available accelerator
  and passes it through to each operator run.

In-kernel profiling
- `kernels/profile.py`: add `has_gpu_timer()`. `time()`/`smid()` need
  `%globaltimer`/`%smid` (CUDA) or `s_memrealtime`/`HW_REG_HW_ID` (HIP); no
  equivalent is currently usable on XPU, so `vector_exp` and `jagged_sum` now
  pass `profile_mem=None` there and compile the instrumentation out rather
  than recording zero timestamps. `jagged_sum`'s `occupancy` metric returns
  None accordingly.

Operator fixes uncovered by the XPU run
- `launch_latency`: `nop_triton_compiled_kernel_run` passed 0 as the stream
  handle. That is the CUDA default-stream sentinel; other backends dereference
  it and crash. Use `driver.active.get_current_stream()`.
- `gdpa`: `torch.cuda.get_device_capability()` raises AssertionError (not
  RuntimeError) on a CUDA-less build, so guard with `torch.cuda.is_available()`.
- `gather_gemv`, `template_attention`: drop the inductor-codegen leftovers
  (`_cuda_getCurrentRawStream`, `empty_strided_cuda`, `torch.cuda._DeviceGuard`)
  in favour of `torch.empty_strided(device=...)` and the device module's guard.

Skip list
- Mark `fused_linear_cross_entropy`, `fused_linear_jsd`, `geglu` and `gdpa` as
  disabled on xpu (CI timeouts) and `template_attention` (IGC crashes on an
  online-softmax `tl.where(m == -inf, ...)` pattern). Each entry documents the
  reason.

Test: `pytest test/test_gpu/main.py` on an Intel Arc B580 -> 42 passed.

+++ tritonbench/operators/grouped_gemm/operator.py
@@ -740,9 +740,13 @@ class Operator(BenchmarkOperator):
             g_lds += [A.stride(0), B.stride(0), C.stride(0)]
 
         # note these are device tensors
-        d_a_ptrs = torch.tensor(A_addrs, device=device)
-        d_b_ptrs = torch.tensor(B_addrs, device=device)
-        d_c_ptrs = torch.tensor(C_addrs, device=device)
+        # Raw data_ptr() values are unsigned 64-bit addresses. On some devices
+        # (e.g. Intel XPU) they exceed INT64_MAX, so torch's default int64
+        # inference raises "Overflow when unpacking long long". Pack them as
+        # uint64 explicitly; the kernel reinterprets each element as a pointer.
+        d_a_ptrs = torch.tensor(A_addrs, dtype=torch.uint64, device=device)
+        d_b_ptrs = torch.tensor(B_addrs, dtype=torch.uint64, device=device)
+        d_c_ptrs = torch.tensor(C_addrs, dtype=torch.uint64, device=device)
         d_g_sizes = torch.tensor(g_sizes, dtype=torch.int32, device=device)
         d_g_lds = torch.tensor(g_lds, dtype=torch.int32, device=device)
